Break up JUnitIntegrationTest into function-specific tests
