+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/junit4/JUnit4TestExecutionIntegrationTest.groovy
@@ -17,9 +17,23 @@
 package org.gradle.testing.junit.junit4
 
 import org.gradle.integtests.fixtures.TargetCoverage
+import org.gradle.integtests.fixtures.TestClassExecutionResult
+import org.gradle.integtests.fixtures.TestExecutionResult
+import org.gradle.testing.junit.AbstractJUnitTestExecutionIntegrationTest
 
 import static org.gradle.testing.fixture.JUnitCoverage.JUNIT_4_LATEST
+import static org.hamcrest.CoreMatchers.containsString
 
 @TargetCoverage({ JUNIT_4_LATEST })
-class JUnit4JUnitIntegrationTest extends AbstractJUnit4JUnitIntegrationTest implements JUnit4MultiVersionTest {
+class JUnit4TestExecutionIntegrationTest extends AbstractJUnitTestExecutionIntegrationTest implements JUnit4MultiVersionTest {
+    @Override
+    String getJUnitVersionAssertion() {
+        return "assertEquals(\"${version}\", new org.junit.runner.JUnitCore().getVersion());"
+    }
+
+    @Override
+    TestClassExecutionResult assertFailedToExecute(TestExecutionResult testResult, String testClassName) {
+        return testResult.testClass(testClassName)
+            .assertTestFailed("initializationError", containsString('ClassFormatError'))
+    }
 }
