+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/junit4/JUnit4TestListenerIntegrationTest.groovy
@@ -14,12 +14,20 @@
  * limitations under the License.
  */
 
-package org.gradle.testing.junit.jupiter
+package org.gradle.testing.junit.junit4
 
 import org.gradle.integtests.fixtures.TargetCoverage
-import org.gradle.testing.fixture.JUnitCoverage
-import org.gradle.testing.junit.AbstractJUnitIntegrationTest
 
-@TargetCoverage({ JUnitCoverage.JUNIT_JUPITER })
-class JUnitJupiterJUnitIntegrationTest extends AbstractJUnitIntegrationTest implements JUnitJupiterMultiVersionTest {
+import static org.gradle.testing.fixture.JUnitCoverage.JUNIT_4_LATEST
+import static org.gradle.testing.fixture.JUnitCoverage.getLATEST_JUNIT3_VERSION
+
+@TargetCoverage({ JUNIT_4_LATEST })
+class JUnit4TestListenerIntegrationTest extends AbstractJUnit4TestListenerIntegrationTest implements JUnit4MultiVersionTest {
+    @Override
+    String getTestFrameworkJUnit3Dependencies() {
+        return """
+            testCompileOnly 'junit:junit:${LATEST_JUNIT3_VERSION}'
+            testRuntimeOnly 'junit:junit:${version}'
+        """
+    }
 }
